pyproject.toml
[project]
name = "mi-coach-brain-backend"
version = "0.1.0"
description = "API for the MiCoachBrain project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=4.2.29",
    "django-environ>=0.11.2",
    "djangorestframework>=3.15.2",
    "strands-agents[gemini]>=1.31.0",
    "strands-agents-tools>=0.2.23",
]
